assignment_1/task_1/arroio_do_sal/solucao_final/requirements.txt
boto3==1.42.83
cryptography==46.0.6
PyMySQL==1.1.2
python-dotenv==1.2.2
